+++ src/main/java/com/chan/stock_portfolio_backtest_api/repository/PortfolioRepository.java
@@ -3,5 +3,5 @@ package com.chan.stock_portfolio_backtest_api.repository;
 import com.chan.stock_portfolio_backtest_api.domain.Portfolio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioRepository extends JpaRepository<Portfolio, Long> {
+public interface PortfolioRepository extends JpaRepository<Portfolio, Integer> {
 }
